+++ gammapy/irf/edisp/tests/test_core.py
@@ -105,15 +105,6 @@ class TestEnergyDispersion2D:
 
 
 @requires_data("gammapy-data")
-def test_get_bias_energy():
-    """Obs read from file"""
-    rmffile = "$GAMMAPY_DATA/joint-crab/spectra/hess/rmf_obs23523.fits"
-    edisp = EDispKernel.read(rmffile)
-    thresh_lo = edisp.get_bias_energy(0.1)
-    assert_allclose(thresh_lo.to("TeV").value, 0.9174, rtol=1e-4)
-
-
-@requires_data("gammapy-data")
 def test_edisp2d_pointlike():
     filename = "$GAMMAPY_DATA/joint-crab/dl3/magic/run_05029748_DL3.fits"
