+++ gammapy/irf/psf/__init__.py
@@ -1,4 +1,4 @@
 from .kernel import *
 from .parametric import *
-from .psf_map import *
+from .map import *
 from .table import *
